Rework GH workflows
There are now two workflows:
1. Test
2. Publish

Test is for branches/pull requests where it'll run the test suite and linting
while publish will build the application and publish it (either to a release
or just upload the artifacts to the action).

Both of these can be dispatched on any arbitrary branch. The goal here is that
we don't really need to build the app on every PR/commit and that doing so
adds significant time to the workflow. This way, we can trigger builds only
when we need it.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>